Combines TabsPagerListener and TabsCreator in to a single class which better encapsulates the responsibility for handling changes to the PagerAdapter data
